fix(admin-toolkit): restore dropped volume/stream-key behaviors; drop dead code and noisy comments
Post-review follow-up:
- Volume slider now spans 0..1 so screens can be muted.
- Extract a shared VolumeSlider (used by Video URL / DCL Cast / Stream) that also restores the 'Sound is disabled for all screens' notice.
- Restore the stream-key security warning under the stream key.
- Delete the dead button-based VolumeControl and the now-unused ICONS/VOLUME_STEP exports.
- Remove redundant JSX section-label comments, commented-out rewards code, and obvious step comments.
